Initial commit
The volume of a sphere is 4/3πr3, where π has the value of "pi" given in Section 2.1 of your textbook. 
Write a function called print_volume (r) that takes an argument for the radius of the sphere, and prints
the volume of the sphere.

Call your print_volume function three times with different values for radius.

Include all of the following in your Learning Journal:

The code for your print_volume function.
The inputs and outputs to three calls of your print_volume.